Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float a;
+    double float a;
     scanf("%.4lf",&a);
     printf("You entered: %.4lf", a);
     return 0;
